Fix skewed image & fix loading data every frame
The preview image was skewed when the image_width * the number of
channels per pixel was not a multiple of 4. This is because OpenGL's
row alignment is by default 4 bytes. A possible fix was to change it
to 0, but I decided it was better if I simply added an alpha channel to
the image, that way it is RGBA instead of RGB. Most CPUs and GPUs like
to with 32 bit (4 byte) chunks, instead of 24 bit (3 byte) chunks, so
the image would get converted in the GPU by the driver anyways.

More information can be found here:
https://www.khronos.org/opengl/wiki/Common_Mistakes
At the "Texture upload and pixel reads" section.

This commit also fixes a bug I noticed while debugging that loads the
pixel buffer into the image every frame instead of only when the buffer
changes.